Breaking out request runner tests
